refactor: extract upload utilities
